backend/requirements.txt
# Family Network Backend
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<5
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator>=2.0.0
python-multipart==0.0.17
